add interfaces, stripe secret, logs, check line`s length
