+++ README.md
@@ -1,8 +1,6 @@
-stripe
-
+stripe credentials:
 revinskaya.katsiaryna@ifortex.com
 123gaw4R121p
 
+to test app follow the link
 http://localhost:8080/checkout
-
-you should add STRIPE_SECRET_KEY
