+++ src/main/java/org/example/config/SecurityConfig.java
@@ -1,7 +1,7 @@
 package org.example.config;
 
-import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
 import org.example.services.CustomUserDetailsService;
+import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationProvider;
