+++ src/main/java/org/example/controllers/CheckoutController.java
@@ -1,8 +1,8 @@
 package org.example.controllers;
 
-import org.example.data.User;
-import org.example.repository.UserRepository;
 import org.example.services.UserService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
@@ -15,9 +15,12 @@ public class CheckoutController {
     @Autowired
     private UserService userService;
 
+    private static final Logger logger = LoggerFactory.getLogger(CheckoutController.class);
+
     @GetMapping("/checkout")
     public String showCheckoutForm(Model model, Authentication authentication) {
         String email = authentication.getName();
+        logger.info("client with email {} asked for stripe checkout page", email);
         boolean hasStripeCustomer = userService.hasStripeCustomer(email);
         model.addAttribute("hasStripeCustomer", hasStripeCustomer);
         return "checkout";
