+++ src/main/java/org/example/controllers/SubscriptionController.java
@@ -5,6 +5,8 @@ import com.stripe.model.billingportal.Session;
 import org.example.data.User;
 import org.example.services.CreateStripeSessionsService;
 import org.example.services.UserService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -22,6 +24,8 @@ public class SubscriptionController {
     @Autowired
     private CreateStripeSessionsService createStripeSessionsService;
 
+    private static final Logger logger = LoggerFactory.getLogger(SubscriptionController.class);
+
     @PostMapping("/create-checkout-session")
     public Map<String, String> createCheckoutSession(
             @RequestBody Map<String, String> payload,
@@ -32,9 +36,9 @@ public class SubscriptionController {
 
         com.stripe.model.checkout.Session session = createStripeSessionsService
                 .createCheckoutStripeSession(userEmail, priceId);
-
         Map<String, String> response = new HashMap<>();
         response.put("url", session.getUrl());
+        logger.info("create checkout session {} for user {}", session.getUrl(), authentication.getName());
         return response;
     }
 
@@ -46,6 +50,7 @@ public class SubscriptionController {
 
         Map<String, String> response = new HashMap<>();
         response.put("url", portalSession.getUrl());
+        logger.info("create customer portal session {} for user {}", portalSession.getUrl(), authentication.getName());
         return response;
     }
 }
