+++ src/main/java/org/example/data/Subscription.java
@@ -3,7 +3,6 @@ package org.example.data;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 
 @Entity
