+++ src/main/java/org/example/repository/UserRepository.java
@@ -7,5 +7,6 @@ import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByStripeCustomerId(String stripeCustomerId);
+
     Optional<User> findByEmail(String email);
 }
