+++ src/main/resources/application.properties
@@ -12,6 +12,7 @@ logging.level.org.hibernate.SQL=DEBUG
 logging.level.org.hibernate.type.descriptor.sql=TRACE
 
 STRIPE_PUBLIC_KEY=pk_test_51RkjgNH13tMkI9AuOa2nAwnMbmPnAGYPHR9dqYUVVixEfVktpQTcMB221dpedAmKpocKQ0x6nvfCr2KITMgZcZ5500iyLjZcXH
+STRIPE_SECRET_KEY=sk_test_51RkjgNH13tMkI9AuMcw5ynYsqVH1SiNBgzP53wJMJ6pwu2wtukfvnai45hsz6fMBUxBzLv7XemCdc1JY8rUGRYdS00cg0z3uS9
 
 stripe.webhook.secret=whsec_yKhTFSae7PZherMdMZX6boS7LGD9HGkD
